Coloring book seeding and global properties window
Application can now place seeds to generate coloring book patterns on lines of symmetry, and knows when to take a screenshot of the page.

There's now also a global properties window, where settings related to the size and spacing of points, the behavior of line growth and decay, as well as coloring book seed generation can all be edited.